evm_transition_tool: refactor each tool to separate files
